use v1 Namespaced api + refactor common code into utils file
